app/requirements.txt
sqlalchemy==2.0.35
sqlmodel==0.0.22
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
psycopg[binary]>=3.2.1 ; extra == "postgres"